weather_bot/requirements.txt
google-adk==0.1.0
litellm==1.30.7
python-dotenv==1.0.1
